Get House Price Data

The UK House Price Index (UK HPI) captures changes in the value of residential properties. The UK HPI uses sales data collected on residential housing transactions, whether for cash or with a mortgage. Data is available at a national and regional level, as well as counties, local authorities and London boroughs.

ukhp_get(
  region = "england",
  item = "housePriceIndex",
  regexp = FALSE,
  start_date = NULL,
  end_date = NULL,
  ...
)

Arguments

region

the region to select. If regexp is set to FALSE then the matching should be exact, see ukhp_avail_regions for available regions. If regexp is set to TRUE then partial matching is possible. Furthermore if it is set to NULL then selects all available regions.

item

the item to select. See ukhp_avail_items for the available categories.

regexp

use regular expression in sparql to search for regions.

start_date

the start date as YYYY-MM-DD.

end_date

the end date as YYYY-MM-DD.

...

query modifiers passed through rdf_modifiers.

Value

Returns a tibble in long format.

Details

Properties have been included:

in England and Wales since January 1995

in Scotland since January 2004

in Northern Ireland since January 2005

See also

Examples

# \donttest{
# This is case sensitive
ukhp_get("england")
#> # A tibble: 691 × 3
#>    region  date       housePriceIndex
#>    <fct>   <date>               <dbl>
#>  1 england 1968-04-01             1.1
#>  2 england 1968-04-01             1.1
#>  3 england 1968-04-01             1.1
#>  4 england 1968-07-01             1.2
#>  5 england 1968-07-01             1.2
#>  6 england 1968-07-01             1.2
#>  7 england 1968-10-01             1.2
#>  8 england 1968-10-01             1.2
#>  9 england 1968-10-01             1.2
#> 10 england 1969-01-01             1.2
#> # ℹ 681 more rows

# However you can use regular expression instead of exact match
ukhp_get("england", regexp = TRUE)
#> # A tibble: 1,464 × 3
#>    region  date       housePriceIndex
#>    <fct>   <date>               <dbl>
#>  1 england 1968-04-01             1.1
#>  2 england 1968-04-01             1.1
#>  3 england 1968-04-01             1.1
#>  4 england 1968-07-01             1.2
#>  5 england 1968-07-01             1.2
#>  6 england 1968-07-01             1.2
#>  7 england 1968-10-01             1.2
#>  8 england 1968-10-01             1.2
#>  9 england 1968-10-01             1.2
#> 10 england 1969-01-01             1.2
#> # ℹ 1,454 more rows

# For all available items
ukhp_avail_items()
#>  [1] "averagePrice"                             
#>  [2] "housePriceIndex"                          
#>  [3] "percentageChange"                         
#>  [4] "percentageAnnualChange"                   
#>  [5] "salesVolume"                              
#>  [6] "averagePriceDetached"                     
#>  [7] "housePriceIndexDetached"                  
#>  [8] "percentageChangeDetached"                 
#>  [9] "percentageAnnualChangeDetached"           
#> [10] "averagePriceSemiDetached"                 
#> [11] "housePriceIndexSemiDetached"              
#> [12] "percentageChangeSemiDetached"             
#> [13] "percentageAnnualChangeSemiDetached"       
#> [14] "averagePriceTerraced"                     
#> [15] "housePriceIndexTerraced"                  
#> [16] "percentageChangeTerraced"                 
#> [17] "percentageAnnualChangeTerraced"           
#> [18] "averagePriceFlatMaisonette"               
#> [19] "housePriceIndexFlatMaisonette"            
#> [20] "percentageChangeFlatMaisonette"           
#> [21] "percentageAnnualChangeFlatMaisonette"     
#> [22] "averagePriceCash"                         
#> [23] "housePriceIndexCash"                      
#> [24] "percentageChangeCash"                     
#> [25] "percentageAnnualChangeCash"               
#> [26] "averagePriceMortgage"                     
#> [27] "housePriceIndexMortgage"                  
#> [28] "percentageChangeMortgage"                 
#> [29] "percentageAnnualChangeMortgage"           
#> [30] "averagePriceFirstTimeBuyer"               
#> [31] "housePriceIndexFirstTimeBuyer"            
#> [32] "percentageChangeFirstTimeBuyer"           
#> [33] "percentageAnnualChangeFirstTimeBuyer"     
#> [34] "averagePriceFormerOwnerOccupier"          
#> [35] "housePriceIndexFormerOwnerOccupier"       
#> [36] "percentageChangeFormerOwnerOccupier"      
#> [37] "percentageAnnualChangeFormerOwnerOccupier"
#> [38] "averagePriceNewBuild"                     
#> [39] "housePriceIndexNewBuild"                  
#> [40] "percentageChangeNewBuild"                 
#> [41] "percentageAnnualChangeNewBuild"           
#> [42] "averagePriceExistingProperty"             
#> [43] "housePriceIndexExistingProperty"          
#> [44] "percentageChangeExistingProperty"         
#> [45] "percentageAnnualChangeExistingProperty"   
#> [46] "salesVolumeExistingProperty"              
#> [47] "salesVolumeNewBuild"                      
#> [48] "salesVolumeCash"                          
#> [49] "salesVolumeMortgage"                      

ukhp_get(c("england", "wales"), item = c("salesVolume", "housePriceIndexDetached"))
#> # A tibble: 736 × 4
#>    region  date       salesVolume housePriceIndexDetached
#>    <fct>   <date>           <dbl>                   <dbl>
#>  1 england 1995-12-01       69345                    18.2
#>  2 england 1995-03-01       67025                    18.2
#>  3 england 1996-02-01       51401                    18.2
#>  4 england 1995-02-01       47880                    18.3
#>  5 england 1996-01-01       48694                    18.3
#>  6 england 1996-03-01       68453                    18.3
#>  7 england 1995-11-01       63955                    18.4
#>  8 england 1995-05-01       64192                    18.4
#>  9 england 1995-06-01       72640                    18.4
#> 10 england 1995-09-01       67251                    18.4
#> # ℹ 726 more rows
# }