Get Transaction Data

Transaction Data is a dataset that shows how many customer applications we completed, in the preceding month for: first registrations, leases, transfers of part, dealings, official copies and searches. This is based on customer and location.

uktrans_get(
  item = "totalApplicationCountByRegion",
  region = NULL,
  regexp = TRUE,
  start_date = NULL,
  end_date = NULL,
  ...
)

Arguments

item
item to select, see uktrans_avail_items() for available items.

region
region to select, see uktrans_avail_regions() for available regions.

regexp
use regular expression in sparql to search for regions.

start_date
the start date as YYYY-MM-DD.

end_date
the end date as YYYY-MM-DD.

...
query modifiers passed through rdf_modifiers.

Value

Returns a tibble in long format.

Examples

# \donttest{
uktrans_get(item = "totalApplicationCountByRegion", region = "East Anglia")

#> # A tibble: 112 x 3
#>    region      date       totalApplicationCountByRegion
#>    <chr>       <date>                             <dbl>
#>  1 East Anglia 2011-12-01                         37819
#>  2 East Anglia 2012-01-01                         44231
#>  3 East Anglia 2012-02-01                         44453
#>  4 East Anglia 2012-03-01                         46814
#>  5 East Anglia 2012-04-01                         40693
#>  6 East Anglia 2012-05-01                         47885
#>  7 East Anglia 2012-06-01                         39506
#>  8 East Anglia 2012-07-01                         46539
#>  9 East Anglia 2012-08-01                         45942
#> 10 East Anglia 2012-09-01                         41976
#> # … with 102 more rows

# If `region` is left as NULL then it returns all available regions
uktrans_get(item = "totalApplicationCountByRegion")

#> # A tibble: 1,344 x 3
#>    region     date       totalApplicationCountByRegion
#>    <chr>      <date>                             <dbl>
#>  1 South East 2011-12-01                        214852
#>  2 South East 2012-01-01                        247768
#>  3 South East 2012-02-01                        250045
#>  4 South East 2012-03-01                        271010
#>  5 South East 2012-04-01                        233486
#>  6 South East 2012-05-01                        267587
#>  7 South East 2012-06-01                        222280
#>  8 South East 2012-07-01                        263214
#>  9 South East 2012-08-01                        255424
#> 10 South East 2012-09-01                        233684
#> # … with 1,334 more rows

# Quering all available transaction data
uktrans_get(item = uktrans_avail_items())

#> # A tibble: 1,344 x 14
#>    region     date       DFLCount DLGCount FRCount TPCount OSPCount OSWCount
#>    <chr>      <date>        <dbl>    <dbl>   <dbl>   <dbl>    <dbl>    <dbl>
#>  1 South East 2011-12-01     2969    74122    1846    2010     5859    29053
#>  2 South East 2012-01-01     3950    71168    2106    2874     4571    28591
#>  3 South East 2012-02-01     4071    63747    2229    3118     4811    28164
#>  4 South East 2012-03-01     4126    70021    2382    2578     6183    32853
#>  5 South East 2012-04-01     3802    65445    1821    2055     4306    25757
#>  6 South East 2012-05-01     3951    69133    1993    2426     5113    32780
#>  7 South East 2012-06-01     2940    58046    1838    2138     5478    28140
#>  8 South East 2012-07-01     3675    73408    2271    2769     4697    33578
#>  9 South East 2012-08-01     2978    72609    1959    2549     5114    32433
#> 10 South East 2012-09-01     2951    63909    1837    2376     4882    28948
#> # … with 1,334 more rows, and 6 more variables: OSNPWCount <dbl>,
#> #   OC1Count <dbl>, SIMCount <dbl>, OSNPPCount <dbl>, OC2Count <dbl>,
#> #   totalApplicationCountByRegion <dbl>

# }